requirements.txt
blinker==1.4
boto3==1.4.0
botocore==1.4.49
cassandra-driver==3.7.1
click==6.6
docutils==0.12
Flask==0.10.1
futures==3.0.5
itsdangerous==0.23
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
MySQL-python==1.2.5
mysqlclient==1.3.7
packaging==16.8
psutil==5.1.0
pyparsing==2.1.10
python-dateutil==2.5.3
python-gcm==0.2
pytz==2016.10
requests==2.10.0
s3transfer==0.1.2
six==1.10.0
Werkzeug==0.9.4
